Jump list feature:
Creates a list of places to jump to based on the current cursor
position. Can be mapped using <Plug>(php_jump_list). This is not
really what I have in mind for this plugin, but it is a starting
point for getting something working with Vim's include-search.
This mapping is borrowed from the Vim help.